Create the baseline app
Enough pieces have been put together to show the problem that should be
solved.  Entering two team names and scores and pressing submit
calculates a winner and loser and then updates the record for each team.
When that happens, the div for each division is being updated instead of
only updating the division that has changed.