separate the draw sections into functions
The draw logic for each division is split into a separate function with
its own subscription call.  Just separating to functions was not enough
to make the updates only happen when that section changed.  The
functions had to be wrapped with square brackets rather than parens
before it all worked.